backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas==2.1.4
tqdm>=4.66.1
playwright>=1.41.2
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
sqlite3-api>=0.1.0
langchain>=0.1.12
langchain-community>=0.0.28
langchain-google-genai>=0.0.11
google-generativeai>=0.3.2
sentence-transformers>=2.5.1
faiss-cpu>=1.7.4
chromadb==0.4.24
pypdf==4.1.0
huggingface-hub>=0.20.3
langchain-huggingface